Updated duplicated code is printed in logfile
Updated the existing code with the addition of finding the longest common substring and then printing it into a log file for inspection. This will lead to better understandability of where the duplication actually occurs. The threshold of how long a common substring can be changed as an argument to the command, furthermore the amount of printed common substrings can also be altered via an argument. The default is the three longest common substrings. For the threshold, it starts at 25 to not include duplicates that are common and not a code smell, such as different keywords.